+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `map<k,v>` support to the Worker Inspector window. [#1403](https://github.com/spatialos/gdk-for-unity/pull/1403)
 - Added `Open inspector V2` menu item that opens the new inspector in the browser. [#1407](https://github.com/spatialos/gdk-for-unity/pull/1407)
 - Added the option to set `UseExternalIp` using the command line argument `+useExternalIp` with the `CommandLineConnectionParameterInitializer`. [#1409](https://github.com/spatialos/gdk-for-unity/pull/1409)
+- Added support for flag arguments in `CommanLineParser`. [#1409](https://github.com/spatialos/gdk-for-unity/pull/1409)
 
 ### Changed